+++ CHANGELOG.md
@@ -1,15 +1,23 @@
-## x.y.z (July 13, 2022)
+## 0.1.0 (July 13, 2022)
 * github release refactor
 * zsec update for terraform support up to 1.1.9
 * zsec updated with mac m1 option for terraform arm64 version download
 * azurerm provider updated to 2.99.0 for all module support and deployment types
 * modules renamed for granularity and consistency
 * azurerm_public_ip resource updated per Network API version 2020-08-01 (https://azure.microsoft.com/en-us/updates/zone-behavior-change/)
-* NSG resources broken out into individual child module
+* NSG resources broken out into individual child module with reuse and byo nsg added
 * zsec enhancement inputs
 * ZS-17339 - support for Managed Identities in different Azure Subsciptions
+* Managed Identity resource broken out into individual child module
 * zsec full service deployment support
+* terraform.tfvars customized per deployment type
+* base_cc renamed to base_1cc
+* bug-124439 - accelerated_networking_enabled variable added to service interfaces. left default false for now until support added
+* workload and bastion hosts changed from ubuntu to centos
 
+
+previous changelog would all disappear
+################################################################
 
 ## 1.1.2 (June 13, 2022)
 NOTES: